apps/desktop/backend/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
requests>=2.28.0
websockets>=11.0.0
python-multipart
pyyaml

# Document processing
python-docx
openpyxl
jinja2

# Vision/OCR
opencv-python
pytesseract

# Optional: Install faiss-cpu and chromadb separately
# See requirements-optional.txt for details
